internal/efi: Add HostEnvironment bits for efi/preinstall
This extends the HostEnvironment interface to accommodate efi/preinstall,
to make it a superset of the one that's already exposed by the efi package.
The intention is that this superset won't be exposed for now, other than
for unit testing - the efi package still only exposes the existing
subset.

The defaultEnvImpl in has been updated to provide a default implementation of
it, and the internal/efitest package has been updated to be able to mock the
new parts of it.