tpm2: Use tpm2.TPMDevice abstraction
Secboot currently opens a transport to the Linux character device at a
hardcoded path (/dev/tpm0) and passes the opened transport to the
deprecated tpm2.NewTPMContext API. The replacement API
(tpm2.OpenTPMDevice) makes use of a new abstraction (tpm2.TPMDevice).

The linux sub-package provides a way of enumerating TPM devices or
providing access to the default one. It also provides access to the
resource managed device, which we might want to use in secboot in the
future, and it provides access to the physical presence interface for
performing actions that are implemented in platform firmware.

Implementing this now is required to write test cases for
https://github.com/canonical/secboot/issues/353, as the current test
harness will permit the code under test to open as many TPM connections
as it likes, when in reality, it can only have a single open connection
if it is using the direct character device (/dev/tpm0).