preinstall: Add checks for virtualization mode
If a container environment is detected, this always return an error.
Detecting a VM returns a value which will be interpreted by the caller
(RunChecks) and will also return an error unless a flag is specified
by snapd to suppress this, in which case the subsequent firmware
protection tests will be skipped because these don't make sense in
virtualized environments.