internal/efitest: revert the change to MockHostEnvironment.VarContext
The previously committed change only added an implementation of
efi.VarsBackend to the supplied context if the supplied
efitest.MockVars was not nil, but this resulted in existing tests having
to explicitly create an empty efitest.MockVars.

Revert to the existing behaviour where we add an implementation of
efi.VarsBackend to the supplied context in all cases. Although the
implementation is a nil map, that is ok for reading (it's only adding
keys to a nil map that causes go to panic).