preinstall: Add simple checks for PCRs 0 and 2
This adds some trivial to checks to both PCRs 0 and 2:
- that they contain events of the expected type.
- that there are no measurements to them from the OS-present phase.

As we don't do any sort of prediction for these PCRs (yet), it doesn't
perform additional checking such as ensuring that the event data is
consistent with the event digests when the event digests are a tagged
hash of the event data (ie, not purely informational).

Note that this modifies the internal/efitest package to emit
EV_S_CRTM_VERSION and EV_EFI_PLATFORM_FIRMWARE_BLOB events with the
correct event data types.