argon2: Add helpers for running the KDF remotely
As Argon2 is memory intensive, it's not suitable for multiple invocations in
long-lived garbage collected processes. For this reason, Argon2 is
abstracted with an interface (Argon2KDF), of which the application sets
a global version of this which is intended to proxy KDF requests to a
short-lived remote process which uses the real InProcessArgon2KDF.

This adds some functionality to facilitate this.

First of all, InProcessArgon2KDF is no longer a variable - it's a
function. By default, it's methods return an error unless the
application code has called SetIsArgon2RemoteProcess, which unlocks the
real in-process KDF.

Then there are JSON serializable types "Argon2RemoteInput" and
"Argon2RemoteOutput". The input can be fed directly to
RunArgon2RequestInRemoteProcess on the remote side, but this is a
fairly low-level API.

There is a higher level API - NewRemoteArgon2KDF, for use in the
application process, and which returns an implementation of Argon2KDF
which proxies requests to a short-lived remote helper process. The
caller supplied a function to construct an appropriate exec.Cmd
instance for this. This function is configured so that the remote
process recieves a request on stdin and it expects a response on
stdout. The remote process passes both os.Stdin and os.Stdout to
WaitAndRunArgon2RequestInRemoteProcess, although it doesn't hardcode
these descriptors for implementations that want to construct their own
transport that doesn't rely on stdin and stdout.

Once a remote process has completed a request, it should exit cleanly.
Neither RunArgon2RequestInRemoteProcess or
WaitAndRunArgon2RequestInRemoteProcess support being called more than
once in the same process.

The code in cmd/run_argon2 provides an example remote process, although
this is mainly useful for unit testing (where it is currently used). It
is envisaged that the remote process will be a special mode of snapd and
snap-bootstrap in order to avoid adding an additional new go binary just
for this.