preinstall: Add minimal checks for PCR 2
This adds a minimal check for PCR 2 which determines whether any
code from value-added-retailer components is executing or not. This
will be used by RunChecks because there will be an option to reject
systems that include code from value-added-retailer components.

As efi.WithDriversAndAppsProfile just copies events from the log and
does no prediction for this PCR, this function doesn't do any more
extensive testing, such as ensuring that the PCR only contains events
of the expected type, the event data for each event is of an expected
type, and for events where the digest is a tagged hash of the event
data, that the digest is consistent with the event data.