preinstall: Add public RunChecks API
RunChecks is the first main entry point for running all of the platform
tests included in this package. It is considered to be quite a low
level and a future PR is going to add a higher level API that is easier
to use (RunChecksContext)

It can be customized with the CheckFlags argument, and the caller must
supply a list of images loaded during the current boot, in the correct
order.

On failure, it will return an error. In some cases, it fails early. In
other cases, it carries on and collects individual errors which are
then returned wrapped in RunChecksError. This can be obtained with the
errors.As API in order to obtain individual errors.

On success, it will return a CheckResult. Any errors that were
encountered but are considered non-fatal based on the supplied
CheckFlags will he added to the Warnings field of CheckResult.

The intention is that CheckResult is persisted and then used by an API
(to be added in the next PR) along with some user customization flags in
order to select the most appropriate set of TCG defined PCRs to seal
against. This can still fail if an appropriately secure set of PCRs
cannot be selected, even if RunChecks succeeded. The future
higher-level RunChecksContext API will make this case easier to handle
and will also help remove the burden of deciding which set of flags
to pass to RunChecks.