preinstall: Add explicit checks for empty PCR banks
Whilst this has no consequence for FDE because we seal against a good
bank, it breaks measured boot as required by remove verifiers, as an
empty PCR bank provides a means for an adversary to spoof any host
platform of their choosing.

We reject systems with empty PCR banks by default, but with an opt-in to
permit it.