preinstall: rename indentLines and other changes
This renames indentLines to makeIndentedListItem to make it clear that
it's used to help printing out lists of things (in this context, mostly
lists of errors).

It also refactors it to use bufio.Scanner.

Some additional test cases are added to test the formatting of
RunChecksErrors, both on its own and embedded as a warning inside of
CheckResult.