tpm2: Avoid attempting to avoid opening duplicate TPM connections
When using NewTPMPassphraseProtectedKey, it is passed an already open
connection. But the core secboot package calls back into the platform
handler to set the user auth value, which opens a second connection.
When using the direct device (/dev/tpm0), this doesn't work.

This updates the test suite to make sure that this scenario is caught
and results in an error. It also fixes the problem by providing the open
TPM connection to the singleton tpm2 platform handler so that it uses
this rather than trying to open a new connection.

This also identified a bug that the core secboot package was not passing
the role value to the platform handler, which only seemed to be caught
by adding tests that created passphrase protected TPM keys with a role.