Fix plumbing of the role via the PlatformKeyDataHandler interface
PR #361 addressed the issue that most tpm2 tests were being performed
with an empty role, which isn't realistic. It turns out that I missed a
few places, particularly in some of the platform tests, that if I had
fixed in the previous PR, would have noticed that the role parameter
isn't being passed through the platform interface (this would get picked
up when changing a password because key validate would fail).

This addresses this, and fixes tests in the top-level package to catch
it as well.

Note that I'm using the same value in most tests - my normal approach to
unit testing is to test each input with at least more than one value,
but the tests in the top-level secboot package and the tpm2 sub-package
are getting a bit complicated with a lot of inconsistencies in their
approach to testing. I have issues #329 and #330 open to refactor the
tests so they are structured in a more consistent way, like the tests
for the recently added plainkey platform are.