argon2: Ensure a panic in the KDF is handled correctly
The gopkg.in/tomb.v2 package does not handle routines that panic very
well - the active routine count will not decrement and it won't put the
tomb into a dying state. In fact, it looks like it makes it impossible
for the tomb to ever reach a dead state.

If the Argon2 crypto package panics, the KDF routine will disappear but
the child process will carry on processing watchdog requests despite the
fact that the parent is never going to get an answer.

We recover panics on the KDF routine and turn them into errors that put
the tomb into a dying state. This begins the teardown of the child
process and halts the processing of watchdog requests. For this reason,
the parent may see 1 of 2 errors - either an error indicating that the
child process exited with a non-zero exit code, or an error indicating a
watchdog timeout. The test added for this handles both cases.

This also introduces a small package - internal/testenv - with a single
function (testenv.IsTestBinary) that can be used in production code and
requires some special linker flags to enable. This permits us to export
a function from the core secboot library to mock an internal function
that should only be accessible in unit tests (and should panic
otherwise).