tpm2: try an alternative approach
Rather than setting a Connection on the platform handler on the tpm2
side and then the calling PlatformKeyDataHandler.ChangeAuthKey for the
first time, try permitting the Connection to be passed directly to
ChangeAuthKey instead.

This adds a new parameter that can be of any type to
KeyWithPassphraseParams. This is passed directly to
PlatformKeyDataHandler.ChangeAuthKey when it is called for the first
time. Plarforms are free to use this new parameter however they like,
but the tpm2 package uses it to pass an open TPM connection. When called
via KeyData.ChangePassphrase, nothing is passed via the new argument and
the tpm2 package will open a new connection.

This avoids race conditions with the previous approach and introducing
new thread safety issues at a time when I'd like to be reducing these
overall.