tpm2: Directly authenticate the role field of the metadata.
The role field currently isn't authenticated directly becuase it's used
to generate the policyRef for the TPM2_PolicyAuthorize command. However,
the polcyRef is stored in the metadata, and so modifications to the role
field of a keyslot metadata will only be detected by calling a method of
SealedKeyData that calls in to keyData.ValidateData. In this sense, it
is authenticated indireectly.

However, we might make security decisions based on the role of a key
used for unlocking, so the current approach isn't sufficient - it needs
to be authenticated directly during unlocking.