Add a flags argument associated with registered PlatformKeyDataHandlers
When adding features to key metadata, we need to be sure that we have
kernels that contain a snap-bootstrap based on a new enough version of
secboot before those features can be used. Given that other requirements
mean that we will be passing information from early boot to snapd, we
may as well pass information about the secboot packages compiled into
snap-bootstrap as part of that information.

In order to do this, I've added a way for each platform that registers a
handler to be able to specify a set of flags. Each platform has 40-bits
which they can use however they choose. The remaining 24-bits may be
used for common flags in the future (eg, we could have a common flag to
allow a platform to indicate that it supports user auth rather than doing
what we do today - we just call into the platform's RecoverKeysWithAuthKey
and ChangeAuthKey methods and expect them to return an error).

The existing platforms are just using the lower bis of the flags to
indicate the platform version / generation for now, but the 40-bits
should be sufficient to allow a platform to define a mix of version /
generation and / or feature flags.

The stateful API for activation will capture these flags for all
registered platforms.