efi/preinstall: Use BootGuard Status (0x13a) MSR for TPM discreteness detection
Previously the TPM manufacturer being INTC (Intel) was used to determine whether
the TPM is provided by the PTT application in ME firmware or a discrete device.

Instead replace this with checking the BootGuard Status (0x13a) MSR. This should
reliably tell host software whether the TPM is discrete.

Since this is a badly documented feature of the platform, if we fail to
read the MSR or it contains a contradictory value, let's just ignore
this and assume the TPM is discrete (which we assume to be lower
security in regards to tracking platform software integrity).

Also define the value of this MSR in the mocked environments where unit
tests require a firmware TPM.

Signed-off-by: Mate Kukri <mate.kukri@canonical.com>