efi/preinstall: Add some more TPM tests
This adds 2 more TPM tests:
- That no hierarchy has an authorization policy set.
- That the TPM is not in failure mode.

It also moves the lockout check until after we've checked that all of
the hierarchies are unowned. This is because it simplifies potential
options for rectifying this failure later.

Also, the hierarchy ownership checks are rolled into a new error
(TPM2OwnedHierarchiesError) which can capture multiple hierarchies in a
single error, which is a bit more useful.