tests: Use the stable channel for the TPM simulator
I'm in the process of updating the TPM simulator to the one based on
v1.83 of the reference library. Although the default reference build
still keeps SHA1 as a supported algorithm, it only enables the SHA256
and SHA384 PCR banks by default, which is likely to break some tests
where we assume that we have SHA1 banks (as we are able to with the
version of the simulator we've been using for the previous ~2 years).

In order to make the transition smoother, I've copied the previous
version of the TPM simulator to stable, and this updates the workflow
to pull in the simulator from latest/stable. I'll probably need to do
some work in secboot in order to get the tests working in the absence
of the SHA1 bank by default, after which, the new simulator can be
promoted to stable for use in secboot testing.

Note that whilst I was making this change, it became apparent that the
tests currently fail against the latest stable go version (1.24 was
released a week ago). I've verified that I get the same failures locally
as well. As a temporary workaround, I've disabled the tests against
stable and hard-coded go 1.23.

See https://github.com/canonical/secboot/issues/383