efi/preinstall: Use BootGuard Status (0x13a) MSR for TPM discreteness detection
Previously the TPM manufacturer being INTC (Intel) was used to determine whether
the TPM is provided by the PTT application in ME firmware or a discrete device.

Instead replace this with checking the BootGuard Status (0x13a) MSR. This should
reliably tell host software whether the TPM is discrete.

Also modify unit tests to define this MSR based on the TPM type.

Signed-off-by: Mate Kukri <mate.kukri@canonical.com>