preinstall: Simplify NoSuitablePCRAlgorithmError
NoSuitablePCRAlgorithm error had 2 fields from which to obtain errors -
a field for errors that affect an entire PCR bank, and a field for
errors that affect a single PCR in a bank.

Simplify this by having a single field for all errors, which are stored
in a map that is keyed by the PCR bank. PCR-specific errors are added to
this field wrapped in the relevant PCR-specific error type instead, so
that the PCR-specific errors are returned for any errors affecting a PCR,
regardless of whether that error occurs during or after the bank
selection.