preinstall: Be consistent about which errors to return immediately from RunChecks
RunChecks returns some errors immediately, whilst other errors are
grouped together and returned wrapped by RunChecksErrors at the end of
the function. However, there are some inconsistencies with the way this
happens. This PR attempts to clean that up. RunChecks will return errors
immediately in the following cases:
- Testing cannot continue with the error (eg, a PCR bank cannot be
  selected, the TPM is in failure more etc).
- The error cannot be rectified (eg, misconfigured BootGuard, not having
  a TPM2 device etc).

Other errors shouldn't immediately halt testing and should be gathered
to be returned together on completion of RunChecks.