preinstall: use a single type for wrapping multiple errors
There are currently 2 error types that wrap multiple errors:
RunChecksErrors and joinError which is used internally. The additional
API (RunChecksContext) that will come in the next PR also wants a way
to wrap multiple errors, so this removes RunChecksErrors in favour of
a single type which isn't public, and a public interface. The error type
is used both internally and for the public API.